requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2023.10
cryptography==42.0.2
deprecation==2.1.0
Django==4.2
django-cachalot==2.6.2
django-celery-results==2.5.1
django-crispy-forms==2.1
django-email-verification==0.3.3
django-mathfilters==1.0.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
inflection==0.5.1
install==1.3.5
kombu==5.3.5
multidict==6.0.5
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
stripe==8.1.0
transliterate==1.10.2
twilio==8.12.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
validators==0.22.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
